requirements.txt
langchain==0.1.0
langchain-community==0.0.12
langchain-core==0.1.10
python-dotenv==1.0.0
pydantic==2.5.3
rich==13.7.0
pytest==7.4.4